pyproject.toml
[project]
name = "oeis-map"
version = "0.1.0"
description = "Semantic map of the Online Encyclopedia of Integer Sequences"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "pyarrow>=14.0",
    "numpy>=1.24",
    "cohere>=5.0",
    "anthropic>=0.40",
    "umap-learn>=0.5.10",
    "toponymy>=0.4",
    "transformers>=4.0",
    "fast-hdbscan>=0.3",
    "evoc",
    "datamapplot>=0.5",
    "glasbey>=0.2",
    "joblib>=1.3",
    "tqdm>=4.60",
    "python-dotenv>=1.0",
    "nest-asyncio>=1.5",
    "scikit-learn>=1.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.14",
    "pre-commit>=4.0",
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
